Conflicts:
	PC/src/uk/ac/ed/inf/sdp2012/group7/MainRunner.java

+++ PC/src/uk/ac/ed/inf/sdp2012/group7/strategy/planning/OppositionPrediction.java
@@ -27,12 +27,12 @@ public class OppositionPrediction {
 	public ArrayList<Point> getDefaultObstacles(){
 		
 		ArrayList<Point> obstacles = new ArrayList<Point>();
-		int nodeInPixels = this.all_static_objects.getNodeInPixels();
+		double nodeInPixels = this.all_static_objects.getNodeInPixels();
 		
 		Point position = this.all_moving_objects.getTheirPosition();
 		//Possible problem with conversion back to Integers here..?
-		for(int x = (position.x - (2*nodeInPixels)); x <= position.x + 2*nodeInPixels; x = x + nodeInPixels){
-			for(int y = (position.y - (2*nodeInPixels)); y <= position.y + 2*nodeInPixels; y = y + nodeInPixels){
+		for(int x = (int)(position.x - (2*nodeInPixels)); x <= position.x + 2*nodeInPixels; x = x + (int)nodeInPixels){
+			for(int y = (int)(position.y - (2*nodeInPixels)); y <= position.y + 2*nodeInPixels; y = y + (int)nodeInPixels){
 				Point p = new Point(x,y);
 				obstacles.add(p);
 			}
